Move ring fitter into fitting file name
