+++ ctapipe/image/muon/ring_fitter.py
